requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-SocketIO==4.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-engineio==3.12.1
python-socketio==4.5.1
pytube3==9.6.4
requests==2.23.0
selenium==3.141.0
six==1.14.0
typing-extensions==3.7.4.2
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.2.1
